make tcph work, but remove the recipe files so that we can add the recipe as submodule in the next commit

+++ scripts/tpch.sh
@@ -1,5 +1,16 @@
 #!/usr/bin/env bash
 
+if [ ! -f build.gradle ];then
+	echo "please start the script from the root directory of the project"
+	exit -1
+fi
+
+mkdir -p build
+cd build
+dbgen -T L -s 1
+
+exit
+# not used anymore since we use the tpch cookbook
 curl "http://www.tpc.org/tpch/spec/tpch_2_14_0.tgz" -o "build/tpch_2_14_0.tgz"
 cd build
 tar xzf tpch*.tgz
@@ -8,5 +19,5 @@ cp ../cookbooks/tpch/templates/default/makefile.suite dbgen/
 cd dbgen
 make -f makefile.suite
 echo "generating: dbgen -T L -s 1 -v"
-dbgen -T L -s 1 -v
+./dbgen -T L -s 1 -v
 mv lineitem.tbl ../
